+++ Sources/BraintreeDropIn/Models/BTDropInResult.m
@@ -16,7 +16,7 @@
 #pragma mark - Prefetch BTDropInResult
 
 + (void)fetchDropInResultForAuthorization:(NSString *)authorization handler:(BTDropInResultFetchHandler)handler {
-    BTUIKPaymentOptionType lastSelectedPaymentOptionType = [[NSUserDefaults standardUserDefaults] integerForKey:@"BT_dropInLastSelectedPaymentMethodType"];
+    BTDropInPaymentMethodType lastSelectedPaymentMethodType = [[NSUserDefaults standardUserDefaults] integerForKey:@"BT_dropInLastSelectedPaymentMethodType"];
     __block BTAPIClient *apiClient = [[BTAPIClient alloc] initWithAuthorization:authorization sendAnalyticsEvent:NO];
     apiClient = [apiClient copyWithSource:apiClient.metadata.source integration:BTClientMetadataIntegrationDropIn2];
 
@@ -25,11 +25,11 @@
             handler(nil, error);
         } else {
             BTDropInResult *result = [BTDropInResult new];
-            if (lastSelectedPaymentOptionType == BTUIKPaymentOptionTypeApplePay) {
-                result.paymentOptionType = lastSelectedPaymentOptionType;
+            if (lastSelectedPaymentMethodType == BTDropInPaymentMethodTypeApplePay) {
+                result.paymentMethodType = lastSelectedPaymentMethodType;
             } else if (paymentMethodNonces != nil && paymentMethodNonces.count > 0) {
                 BTPaymentMethodNonce *paymentMethod = paymentMethodNonces.firstObject;
-                result.paymentOptionType = [BTUIKViewUtil paymentOptionTypeForPaymentInfoType:paymentMethod.type];
+                result.paymentMethodType = [BTUIKViewUtil paymentMethodTypeForPaymentInfoType:paymentMethod.type];
                 result.paymentMethod = paymentMethod;
             }
             handler(result, error);
@@ -40,12 +40,12 @@
 
 - (UIView *)paymentIcon {
     BTUIKPaymentOptionCardView *paymentOptionCardView = [BTUIKPaymentOptionCardView new];
-    paymentOptionCardView.paymentOptionType = self.paymentOptionType;
+    paymentOptionCardView.paymentMethodType = self.paymentMethodType;
     return paymentOptionCardView;
 }
 
 - (NSString *)paymentDescription {
-    if (self.paymentOptionType == BTUIKPaymentOptionTypeApplePay) {
+    if (self.paymentMethodType == BTDropInPaymentMethodTypeApplePay) {
         return @"Apple Pay";
     } else {
         return self.paymentMethod.paymentDescription;